requirements.txt

Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1
SQLAlchemy==2.0.31
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
xhtml2pdf==0.2.15
python-slugify==8.0.4
